modal functioning css needs to be cleaned up and refined
