+++ app/components/addDropdown.tsx
@@ -1,8 +1,10 @@
 "use client";
 import { Fragment } from "react";
 import { Menu, Transition } from "@headlessui/react";
+import Link from "next/link";
 
 export default function AddDropdown() {
+
 	const dropdownSelectedItemClass = (isActive: boolean) =>
 		isActive
 			? "px-4 py-2 block border-b-2 hover:border-gray-200 text-sm"
@@ -52,9 +54,14 @@ export default function AddDropdown() {
 					<div className="py-1">
 						<Menu.Item>
 							{({ active }) => (
-								<span className={dropdownSelectedItemClass(active)}>
-									Assignment
-								</span>
+								<>
+									<Link
+										href={"?modal=true"}
+										className={dropdownSelectedItemClass(active)}
+									>
+										Assignment
+									</Link>
+								</>
 							)}
 						</Menu.Item>
 						<Menu.Item>
