+++ app/layout.tsx
@@ -4,28 +4,29 @@ import "./globals.css";
 import Navbar from "./components/navbar";
 import ActionBar from "./components/actionbar";
 import WeekDisplay from "./components/weekDisplay";
-import AddAssignment from "./components/addAssignment";
+import AddAssignment from "./components/addAssignmentModal";
 
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
-  title: "Staff Plan",
-  description: "Staff Plan App",
+	title: "Staff Plan",
+	description: "Staff Plan App",
 };
 
 export default function RootLayout({
-  children,
+	children,
 }: Readonly<{
-  children: React.ReactNode;
+	children: React.ReactNode;
 }>) {
-  return (
-    <html lang="en">
-      <body className={inter.className}>
-        <Navbar />
-        <ActionBar />
-        <WeekDisplay />
-        {children}
-      </body>
-    </html>
-  );
+	return (
+		<html lang="en">
+			<body className={inter.className}>
+				<Navbar />
+				<ActionBar />
+				<WeekDisplay />
+        <AddAssignment/>
+				{children}
+			</body>
+		</html>
+	);
 }
